Add Python implementation for asyncio.Future
I'm starting with an explanation of Python futures in Python, so we can
expand on it to build a Rust awaitable that drives a Rust future.

Is this making any sense? The fact that the two languages use the same
denomination is a good sign for compatibility but makes the whole deal a
bit confusing.